Add tree test glob to npm test and move tree slice tests
